String to Int64 validation with luhn algorithm
On a device such as iPhone 4s, Int is a 32-bit integer.
64-bit integer needs to be declared explicitly with Int64.

+++ CCValidator/Classes/CCValidator.swift
@@ -216,7 +216,7 @@ public extension CCValidator {
     }
     
     private class func validateWithLuhnAlgorithm(creditCardNumber number: String) -> Bool {
-        guard let _ = Int(number) else {
+        guard let _ = Int64(number) else {
             //if string is not convertible to int, return false
             return false
         }
